Renamed folders to singular
