+++ src/main/java/vu/psk/ugems/entity/Profile.java
@@ -1,4 +1,4 @@
-package vu.psk.ugems.entities;
+package vu.psk.ugems.entity;
 
 import jakarta.persistence.*;
 import lombok.AllArgsConstructor;
