+++ src/main/java/vu/psk/ugems/repository/GroupRepository.java
@@ -1,8 +1,8 @@
-package vu.psk.ugems.repositories;
+package vu.psk.ugems.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import vu.psk.ugems.entities.Group;
+import vu.psk.ugems.entity.Group;
 
 @Repository
 public interface GroupRepository extends CrudRepository<Group, Long> {
