+++ src/main/java/vu/psk/ugems/repository/ProfileRepository.java
@@ -1,8 +1,8 @@
-package vu.psk.ugems.repositories;
+package vu.psk.ugems.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import vu.psk.ugems.entities.Profile;
+import vu.psk.ugems.entity.Profile;
 
 @Repository
 public interface ProfileRepository extends CrudRepository<Profile, Long> {
