+++ src/main/java/vu/psk/ugems/repository/TaskRepository.java
@@ -1,8 +1,8 @@
-package vu.psk.ugems.repositories;
+package vu.psk.ugems.repository;
 
 import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
-import vu.psk.ugems.entities.Task;
+import vu.psk.ugems.entity.Task;
 
 @Repository
 public interface TaskRepository extends CrudRepository<Task, Long> {
